Refactoring action parameter binding
