--- framework/binders/ParameterTypeFactoryInterface.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace yii\binders;
-
-use ReflectionParameter;
-
-interface ParameterTypeFactoryInterace
-{
-    /**
-     * @param ReflectionParameter $type
-     * @return bool
-     */
-    public function canCreateType($type);
-
-    /**
-     * @param ReflectionParameter $type
-     * @param BindingContext $context
-     * @return mixed
-     */
-    public function createType($type, $context);
-}
