--- framework/binders/RequestBinderInderface.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace yii\binders;
-
-use yii\web\Request;
-
-interface RequestBinderInterface
-{
-    /**
-     * @param BindingContext $context
-     * @return void
-     */
-    public function bind($context);
-}
