+++ framework/bindings/ActionParameterBinder.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\bindings;
+
+use yii\base\BaseObject;
+use yii\base\InlineAction;
+
+class ActionParameterBinder extends BaseObject implements ActionParameterBinderInterface {
+
+    public function getBindingRegistry() {
+        return new BindingRegistry();
+    }
+
+    public function bindActionParams($action, $params)
+    {
+        if ($action instanceof InlineAction) {
+            $method = new \ReflectionMethod($action->controller, $action->actionMethod);
+        } else {
+            $method = new \ReflectionMethod($action, 'run');
+        }
+
+        $bindingRegistry = $this->getBindingRegistry();
+        $bindingContext = new BindingContext(\Yii::$app->request,$action, $params);
+
+        $arguments = [];
+
+        $methodParameters = [];
+        foreach ($method->getParameters() as $param) {
+            $methodParameters[$param->getName()] = $param;
+        }
+
+        foreach ($methodParameters as $name => $param) {
+            $result = $bindingRegistry->create($param, $bindingContext);
+            if ($result instanceof BindingResult) {
+                $arguments[$name] = $result->value;
+            }
+        }
+
+        foreach($arguments as $name => $argument) {
+            if ($argument instanceof ParameterBinderInterface) {
+                $param = $methodParameters[$name];
+                $argument->bindModel($param,  $bindingContext);
+            }
+        }
+
+        return $arguments;
+    }
+}
