+++ framework/bindings/ActionParameterBinderInterface.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\bindings;
+
+interface ActionParameterBinderInterface {
+    /**
+     * @param Action $action
+     * @param array $params
+     * @return array
+     */
+    public function bindActionParams($action, $params);
+}
