+++ framework/bindings/BindingResult.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\bindings;
+
+class BindingResult {
+    /**
+     * @var mixed $value
+     */
+    public $value;
+
+    public function __construct($value)
+    {
+        $this->value;
+    }
+}
